feat(blog): add blog array as a separated component and add it to page
